Update 2.0 fix bugs and add new files.
