+++ main.cpp
@@ -1,0 +1,10 @@
+#include "Functions.h"
+
+using namespace std;
+
+int main() {
+	Functions func;
+	func.mainMenu();
+
+	return 0;
+}
